requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0

# Machine Learning models
xgboost>=1.5.0
lightgbm>=3.2.0

# Data processing
imbalanced-learn>=0.8.0
feature-engine>=1.2.0

# Visualization
plotly>=5.0.0
bokeh>=2.4.0

# Experiment tracking
swanlab>=0.3.0
wandb>=0.12.0

# Configuration
pyyaml>=6.0
hydra-core>=1.1.0

# Utilities
joblib>=1.1.0
psutil>=5.8.0
pathlib2>=2.3.0

# Development
pytest>=6.2.0
pytest-cov>=3.0.0
black>=21.0.0
flake8>=4.0.0
mypy>=0.910

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Optional: Deep Learning (uncomment if needed)
# torch>=1.10.0
# tensorflow>=2.7.0
# pytorch-lightning>=1.5.0